+++ libraries/engage/push-opt-in/selectors/optInTrigger.js
@@ -3,4 +3,4 @@
  * @param {Object} state The current state.
  * @returns {Object} The push opt in information.
  */
-export const getPushOptInTriggerState = state => state?.push?.optInTrigger || {};
+export const getPushOptInTriggerState = state => state?.pushOptIn?.optInTrigger || {};
