+++ themes/theme-ios11/pages/reducers.js
@@ -13,7 +13,7 @@ import menu from '@shopgate/pwa-common/reducers/menu';
 import modal from '@shopgate/pwa-common/reducers/modal';
 import cart from '@shopgate/pwa-common-commerce/cart/reducers';
 import appRating from '@shopgate/engage/app-rating/reducers';
-import push from '@shopgate/engage/push/reducers';
+import pushOptIn from '@shopgate/engage/push-opt-in/reducers';
 import category from '@shopgate/pwa-common-commerce/category/reducers';
 import favorites from '@shopgate/pwa-common-commerce/favorites/reducers';
 import filter from '@shopgate/pwa-common-commerce/filter/reducers';
@@ -63,7 +63,7 @@ const reducers = combineReducers({
   url,
   user,
   appRating,
-  push,
+  pushOptIn,
 });
 
 export default reducers;
