+++ themes/theme-ios11/pages/subscribers.js
@@ -45,7 +45,7 @@ import appConfig from '@shopgate/pwa-common/helpers/config';
 import extensions from 'Extensions/subscribers';
 // App rating
 import appRating from '@shopgate/engage/app-rating/subscriptions';
-import push from '@shopgate/engage/push/subscriptions';
+import pushOptIn from '@shopgate/engage/push-opt-in/subscriptions';
 
 const subscriptions = [
   // Common subscribers.
@@ -77,7 +77,7 @@ const subscriptions = [
   commerceScanner,
   // App rating subscribers
   appRating,
-  push,
+  pushOptIn,
   // Theme subscribers.
   app,
   viewport,
